Pull Error => Status routines into libzip/error
